Split data handling in other file. Now model training only in model_training.py
